requirements.txt
# Core
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.20
Flask-Migrate==4.0.5
alembic==1.12.0

# Authentication & Security
Flask-Login==0.6.2
email-validator==2.0.0.post2
Werkzeug==2.3.7
itsdangerous==2.1.2

# Forms
Flask-WTF==1.1.1
WTForms==3.0.1

# Development
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
